% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_AnnotateMutations.R
\name{AnnotateMutations}
\alias{AnnotateMutations}
\title{Annotate single nucleotide substitutions}
\usage{
AnnotateMutations(
  vcf,
  ref_df = MG1655_ref,
  genome_map = MG1655_genome_map,
  .group = NULL
)
}
\arguments{
\item{vcf}{A tibble with all the entries from vcf file. POS, REF, ALT and INFO columns are
required. Can be generated with ReadVCF function.}

\item{ref_df}{A tibble storing the genomic information of the reference genome. Can be generated
with ExtractGBK function.}

\item{genome_map}{A feature map list to quick trace the feature ids for given positions.
Can be generated using GenerateGenomeMapping function.}

\item{.group}{A character string. The column(s) that should be used to group nucleodides together
when annotating mutations. When annotating multiple VCFs, combine them into one file, while
saving the sample id as an extra column. Use this column name as .group variable so mutations
inside each group are annotated together. If no groupings, use NULL.}
}
\value{
A tibble with annotated mutations. Combine entries in locus_tag with SUB columns in
the GetEA function to get the EA of that mutation.
}
\description{
Read in a vcf tibble then single nucleotide mutations are kept. Mutations
are then annotated according to the reference (gbk.tibble) genome.
}
\author{
Chen Wang
}
