% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_RandomMut.R
\name{RandomMut}
\alias{RandomMut}
\title{RandomMut}
\usage{
RandomMut(
  ref_table,
  DNA_seq,
  ID = "Genome",
  n,
  ti = n\%/\%2,
  cds_only = TRUE,
  replace = FALSE
)
}
\arguments{
\item{ref_table}{The reference table for the reference genome. Should contain these columns:
start, end. Each row represents one gene/intergene.}

\item{DNA_seq}{The DNA seq of the reference genome. Stored as vector, so the nucleotide
at position i can be indexed as DNA_seq[i].}

\item{n}{Number of simulated mutations}

\item{ti}{Number of transition mutations in the simulation. A integer between 0 and n.}

\item{cds_only}{If is TRUE, only mutations in the cds regions will be simulated. Default is TRUE.}

\item{replace}{True or False. Whether a same nucleotide position can be mutated multiple times. Set to FALSE by default.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Generate random mutations on the reference genome and output vcf.
tibble file. The accession ID from Genbank file. Is used to generate #CHROM column in the
vcf file.
}
