% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ReadVCF.R
\name{ReadVCF}
\alias{ReadVCF}
\title{ReadVCF}
\usage{
ReadVCF(path, header_max = 100)
}
\arguments{
\item{path}{Path to the vcf file}

\item{header_max}{Max number of rows to guess VCF header lines.}
}
\value{
A tibble containing all the entries in the VCF file.
}
\description{
Read the entries from the vcf files, headers are removed.
}
\author{
Chen Wang
}
