% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ReadVCF.R
\name{ReadVCF}
\alias{ReadVCF}
\title{ReadVCF}
\usage{
ReadVCF(path)
}
\arguments{
\item{path}{Path to the vcf file}
}
\value{
A tibble containing all the entries in the VCF file.
}
\description{
Read the entries from the vcf files, headers are removed.
}
\author{
Chen Wang
}
